+++ docs/CHANGELOG.md
@@ -3,6 +3,14 @@ Doom Struct (C) 2015-2019
 by Matt Tropiano et al. (see AUTHORS.txt)
 
 
+Slated for 2.1.0
+----------------
+
+- `Added` WadEntry.withNewXXX()
+- `Fixed` WadBuffer.deleteEntry(...) now adjusts offsets properly.
+- `Changed` Reduced visibility of some methods of WadEntry and it is no longer a BinaryObject.
+
+
 Changed in 2.0.0
 ----------------